# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: Сервис управления банковскими картами и переводами
  contact:
    name: Support
    email: cetjs2@vk.com
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: Карты
  description: Управление банковскими картами и операциями
- name: Аутентификация
  description: Методы для получения доступа к системе
paths:
  /cards/transfer:
    post:
      tags:
      - Карты
      summary: Перевод денег
      description: Между картами текущего пользователя
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SucessActionResponse"
      security:
      - bearerAuth: []
  /cards/sendblockrequest/{cardId}:
    post:
      tags:
      - Карты
      summary: Подать заявку на блокировку
      description: "Создает запрос, который должен одобрить админ"
      operationId: sendBlockRequest
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SucessActionResponse"
      security:
      - bearerAuth: []
  /cards/new:
    post:
      tags:
      - Карты
      summary: Выпуск новой карты (Админ)
      description: Доступно только пользователям с ROLE_ADMIN
      operationId: newCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardViewResponse"
      security:
      - bearerAuth: []
  /auth/login:
    post:
      tags:
      - Аутентификация
      summary: Вход в систему
      description: Проверяет логин/пароль и возвращает JWT-токен для доступа к защищенным
        API
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          description: Успешная авторизация
          content:
            '*/*':
              schema:
                type: string
                example:
                  token: eyJhbGci...
        "401":
          description: Неверный логин или пароль
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cards/block/{id}:
    patch:
      tags:
      - Карты
      summary: Заблокировать карту (Админ)
      operationId: blockCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SucessActionResponse"
      security:
      - bearerAuth: []
  /cards/view/{id}:
    get:
      tags:
      - Карты
      summary: Данные о карте
      operationId: getCardInfo
      parameters:
      - name: id
        in: path
        description: ID карты
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardViewResponse"
      security:
      - bearerAuth: []
  /cards/list:
    get:
      tags:
      - Карты
      summary: Список доступных карт
      description: Админ видит всех или фильтрует по пользователю. Юзер видит только
        свои.
      operationId: getAll
      parameters:
      - name: user
        in: query
        description: Логин (только для админа)
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardViewResponse"
      security:
      - bearerAuth: []
  /cards/delete/{id}:
    delete:
      tags:
      - Карты
      summary: Удалить карту (Админ)
      operationId: deleteCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SucessActionResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    TransferRequest:
      type: object
      description: Запрос на перевод денежных средств между картами
      properties:
        fromCardId:
          type: string
          format: uuid
          description: ID карты отправителя
          example: 550e8400-e29b-41d4-a716-446655440000
        toCardId:
          type: string
          format: uuid
          description: ID карты получателя
          example: a1b2c3d4-e5f6-7g8h-9i0j-k1l2m3n4o5p6
        sum:
          type: number
          description: Сумма перевода
          example: 500.0
          minimum: 0.01
      required:
      - fromCardId
      - sum
      - toCardId
    SucessActionResponse:
      type: object
      description: Рростой ответ при успешном выполнении действия
      properties:
        message:
          type: string
          description: Текстовое сообщение об успешном результате операции
          example: Карта 550e8400-e29b-41d4-a716-446655440000 успешно заблокирована
    CreateCardRequest:
      type: object
      description: Данные для создания новой банковской карты
      properties:
        ownerUsername:
          type: string
          description: Уникальное имя пользователя-владельца
          example: ivan_gold
          minLength: 1
        initialBalance:
          type: number
          description: Начальный баланс карты при открытии
          example: 1000.0
        expirationDate:
          type: string
          description: Срок действия карты в формате ММ/ГГ
          example: 12/28
          pattern: MM/yy
      required:
      - expirationDate
      - ownerUsername
    CardViewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cardNumber:
          type: string
        ownerName:
          type: string
        cardStatus:
          type: string
          description: Статус банковской карты
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        expirationDate:
          type: string
        balance:
          type: number
    AuthRequest:
      type: object
      description: Запрос на аутентификацию пользователя
      properties:
        username:
          type: string
          description: Имя пользователя (логин)
          example: ivan_gold
          minLength: 1
        password:
          type: string
          format: password
          description: Пароль пользователя
          example: P@ssw0rd123
          minLength: 1
      required:
      - password
      - username
    ErrorResponse:
      type: object
      description: Объект с описанием ошибки
      properties:
        status:
          type: integer
          format: int32
          description: HTTP статус-код ошибки
          example: 404
        message:
          type: string
          description: Детальное сообщение об ошибке для разработчика
          example: Карта с ID 550e8400-e29b-41d4-a716-446655440000 не найдена
        timestamp:
          type: string
          format: date-time
          description: Точное время возникновения ошибки
    PageCardViewResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardViewResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      description: "Введите JWT токен, полученный в методе /auth/login"
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT